Don't explode if the logic is wrong.
This silently explodes, which is bad, but on the other hand
I have no idea why this 3rd party package is trying to parse
the `sys.argv` when so many other utilities need it.

Without this, I get errors when I pass `pytest -v`.

So, better to just disable this for now.